feat(zk-payment-verifier): extract nullifier registry into its own module
The double-proving prevention this issue asks for was already fully
functional: verify_and_record already rejected duplicate nullifiers,
the nullifier was already a one-way SHA-256 hash revealing nothing
about the payment, and storage already grew as a single linearly-
growing Map (acceptable for MVP per this issue's own acceptance
criteria) -- all three acceptance criteria already had passing tests
(test_nullifier_prevents_double_proof, test_is_nullifier_used,
test_different_services_independent_nullifiers).

What was missing was the dedicated nullifier.rs module the issue's
title and "Affected Code" section describe: the logic lived inline in
lib.rs (storage manipulation) and commitment.rs (compute_nullifier).
Extracted both into nullifier.rs as a self-contained registry API
(compute_nullifier, is_used, record), with lib.rs's contract functions
now just delegating to it. Pure mechanical extraction -- behavior is
unchanged, only the organization changed.

(verifier.rs in this same crate implements a different, more advanced
ZK proof scheme but isn't declared as a module or referenced anywhere
-- pre-existing and out of scope for this issue, left untouched.)

Verified via `rustc --emit=metadata` partial-resolution checks (no
network access to fetch soroban-sdk in this sandbox): identical
pre-existing error profile before/after (just one more expected
"unresolved import soroban_sdk" from the new file importing it), no
new error categories.
