+++ contracts/contracts/zk-payment-verifier/src/commitment.rs
@@ -19,17 +19,3 @@ pub fn compute_commitment(
 
     env.crypto().sha256(&payload).into()
 }
-
-/// Compute nullifier: SHA-256("syncro:payment:v1" || blinding_factor || service_id)
-/// Deterministic per (blinding_factor, service_id) pair — prevents double-proving.
-pub fn compute_nullifier(
-    env: &Env,
-    blinding_factor: &BytesN<32>,
-    service_id: &Bytes,
-) -> BytesN<32> {
-    let mut payload = Bytes::from_slice(env, b"syncro:payment:v1");
-    payload.append(&Bytes::from_slice(env, &blinding_factor.to_array()));
-    payload.append(&service_id.clone());
-
-    env.crypto().sha256(&payload).into()
-}
